refactor: move drop index util to db utils file
