+++ frappe/database/utils.py
@@ -183,3 +183,20 @@ def commit_after_response(func):
 		request.after_response.add(callback_manager.run)
 
 	callback_manager.add(func)
+
+
+def drop_index_if_exists(table: str, index: str):
+	import click
+
+	if not frappe.db.has_index(table, index):
+		click.echo(f"- Skipped {index} index for {table} because it doesn't exist")
+		return
+
+	try:
+		frappe.db.sql_ddl(f"ALTER TABLE `{table}` DROP INDEX `{index}`")
+	except Exception as e:
+		frappe.log_error("Failed to drop index")
+		click.secho(f"x Failed to drop index {index} from {table}\n {e!s}", fg="red")
+		return
+
+	click.echo(f"✓ dropped {index} index from {table}")
